+++ src/JsonApi/Serializer/ItemNormalizerTrait.php
@@ -83,7 +83,7 @@ trait ItemNormalizerTrait
 
         // Surface the client-generated id so the entity setter receives it.
         if ($isPostOperation && $allowClientGeneratedId && isset($data['data']['id'])) {
-            $dataToDenormalize['id'] = $data['data']['id'];
+            $dataToDenormalize[ReservedAttributeNameConverter::JSON_API_RESERVED_ATTRIBUTES['id']] = $data['data']['id'];
         }
 
         return parent::denormalize($dataToDenormalize, $type, $format, $context);